refactor(client): split chat_input_bar into parts
1998 LOC parent down to 465. Eight sibling part files extracted
(attachments, build_helpers, draft_persistence, edit_mode,
keyboard_handling, media_picker, recording_handling, send_handling)
sharing the parent's class scope via part-of. Behaviour-preserving;
flutter analyze clean and all 1921 client tests pass.
