+++ apps/client/lib/src/widgets/chat_input_bar.dart
@@ -53,6 +53,15 @@ import 'input/reply_preview_bar.dart';
 import 'media_picker_panel.dart';
 import 'mobile_media_picker_panel.dart';
 
+part 'chat_input_bar/parts/attachments.dart';
+part 'chat_input_bar/parts/build_helpers.dart';
+part 'chat_input_bar/parts/draft_persistence.dart';
+part 'chat_input_bar/parts/edit_mode.dart';
+part 'chat_input_bar/parts/keyboard_handling.dart';
+part 'chat_input_bar/parts/media_picker.dart';
+part 'chat_input_bar/parts/recording_handling.dart';
+part 'chat_input_bar/parts/send_handling.dart';
+
 /// Extracted chat input bar from ChatPanel.
 ///
 /// Manages:
@@ -65,6 +74,13 @@ import 'mobile_media_picker_panel.dart';
 ///
 /// Exposes [enterEditMode] as a public method so the parent can invoke it
 /// via `GlobalKey<ChatInputBarState>`.
+///
+/// The state class is split across `chat_input_bar/parts/*.dart` files via
+/// `part of` declarations. The parent file owns the widget, the public API,
+/// the state fields, lifecycle, and `build()`. Each part file contains a
+/// library-private extension on [ChatInputBarState] that groups one slice
+/// of behaviour (drafts, send, attachments, edit, recording, media picker,
+/// keyboard, and build helpers).
 class ChatInputBar extends ConsumerStatefulWidget {
   final Conversation conversation;
   final String? selectedTextChannelId;
@@ -168,15 +184,6 @@ class ChatInputBarState extends ConsumerState<ChatInputBar> {
   // Suppresses draft saves during cancel-edit to avoid race with _loadDraft.
   bool _suppressDraftSave = false;
 
-  @override
-  void initState() {
-    super.initState();
-    _controller = ChatInputController();
-    _messageController.addListener(_onTextChanged);
-    _mentionController.addListener(_onMentionChanged);
-    _loadDraft(widget.conversation.id);
-  }
-
   /// Index of the selected row in the mention picker. Reset to 0 each time
   /// the picker opens or the query changes so the first match is always
   /// the keyboard-accept target.
@@ -188,52 +195,13 @@ class ChatInputBarState extends ConsumerState<ChatInputBar> {
   // list each time (audit perf + frontend findings).
   List<String> _cachedMentionCandidates = const [];
 
-  void _onMentionChanged() {
-    if (!mounted) return;
-    final showing = _mentionController.showPicker;
-    final query = _mentionController.query;
-    final changed =
-        showing != _lastMentionShowing || query != _lastMentionQuery;
-    if (!changed) return; // skip rebuild when nothing visible changed
-    _mentionPickerIndex = 0;
-    _cachedMentionCandidates = showing
-        ? MentionAutocomplete.candidateValues(_filteredMentionMembers, query)
-        : const [];
-    _lastMentionShowing = showing;
-    _lastMentionQuery = query;
-    setState(() {});
-  }
-
-  /// Candidate values currently rendered by the mention picker. Cached
-  /// in [_cachedMentionCandidates] and updated only in [_onMentionChanged].
-  List<String> get _mentionCandidates => _cachedMentionCandidates;
-
-  /// Accept the currently-selected mention candidate (Tab/Enter pressed
-  /// while the picker is open). No-op when there are no candidates.
-  void _acceptMentionSelection() {
-    final candidates = _mentionCandidates;
-    if (candidates.isEmpty) return;
-    final idx = _mentionPickerIndex.clamp(0, candidates.length - 1);
-    _handleMentionSelected(candidates[idx]);
-  }
-
-  /// Semantic direction for picker navigation. The candidate list is
-  /// rendered with `reverse: true`, so the *visual* "down arrow"
-  /// actually moves to a smaller index. Encapsulating that here means
-  /// the key handler reads as `_MentionMove.down` rather than
-  /// `_moveMentionSelection(-1)` and a future maintainer can't
-  /// accidentally flip the sign (TD-22).
-  void _moveMentionSelection(_MentionMove direction) {
-    final candidates = _mentionCandidates;
-    if (candidates.isEmpty) return;
-    final delta = switch (direction) {
-      _MentionMove.up => 1, // up arrow → higher index in the reverse list
-      _MentionMove.down => -1, // down arrow → lower index
-    };
-    final next = (_mentionPickerIndex + delta) % candidates.length;
-    setState(() {
-      _mentionPickerIndex = next < 0 ? next + candidates.length : next;
-    });
+  @override
+  void initState() {
+    super.initState();
+    _controller = ChatInputController();
+    _messageController.addListener(_onTextChanged);
+    _mentionController.addListener(_onMentionChanged);
+    _loadDraft(widget.conversation.id);
   }
 
   @override
@@ -345,1421 +313,12 @@ class ChatInputBarState extends ConsumerState<ChatInputBar> {
     );
   }
 
-  // ---------------------------------------------------------------------------
-  // Draft auto-save
-  // ---------------------------------------------------------------------------
-
-  Future<void> _loadDraft(String conversationId) async {
-    final prefs = await SharedPreferences.getInstance();
-    final key = '$_draftKeyPrefix$conversationId';
-    final draft = prefs.getString(key);
-    if (draft != null && draft.isNotEmpty && mounted && !_isEditing) {
-      _messageController.text = draft;
-      setState(() => _isTextEmpty = draft.trim().isEmpty);
-    }
-  }
-
-  void _scheduleDraftSave(String text) {
-    _draftSaveTimer?.cancel();
-    _draftSaveTimer = Timer(const Duration(milliseconds: 600), () {
-      _saveDraftImmediate(widget.conversation.id, text);
-    });
-  }
-
-  Future<void> _saveDraftImmediate(String conversationId, String text) async {
-    final prefs = await SharedPreferences.getInstance();
-    final key = '$_draftKeyPrefix$conversationId';
-    if (text.isEmpty) {
-      await prefs.remove(key);
-    } else {
-      await prefs.setString(key, text);
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Text listener
-  // ---------------------------------------------------------------------------
-
-  void _onTextChanged() {
-    final text = _messageController.text;
-    final empty = text.trim().isEmpty;
-    // Detect multiline either via explicit newlines or content long enough
-    // to soft-wrap at the current viewport width. The character threshold
-    // is recomputed in build() from MediaQuery so a 4K-wide input doesn't
-    // jump to multi-line halfway through a normal sentence.
-    final multiline =
-        text.contains('\n') || text.length > _multilineCharThreshold;
-    if (empty != _isTextEmpty || multiline != _isMultiline) {
-      setState(() {
-        _isTextEmpty = empty;
-        _isMultiline = multiline;
-      });
-    }
-    // Schedule draft save when not in edit mode and not suppressed
-    if (!_isEditing && !_suppressDraftSave) {
-      _scheduleDraftSave(text);
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Send message
-  // ---------------------------------------------------------------------------
-
-  Future<void> _sendMessage() async {
-    final caption = _messageController.text.trim();
-
-    // Attachments still uploading -- don't send text-only and lose them.
-    if (_hasPendingAttachment && !_allPendingAttachmentsReady) {
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Attachments still uploading...',
-          type: ToastType.info,
-        );
-      }
-      return;
-    }
-
-    // If there are uploaded attachments, send them as separate messages.
-    if (_hasPendingAttachment && _allPendingAttachmentsReady) {
-      await _sendAttachments(caption);
-      return;
-    }
-
-    // (legacy guard kept; the above already covers the in-flight case)
-    if (_isAnyPendingAttachmentUploading) {
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Attachment still uploading...',
-          type: ToastType.info,
-        );
-      }
-      return;
-    }
-
-    final text = caption;
-    if (text.isEmpty) return;
-
-    // Slash-command interception: parse before encrypting/sending.
-    final handled = await _tryHandleSlashCommand(text);
-    if (handled) return;
-
-    await _doSend(text);
-    _clearInputAndNotify();
-  }
-
-  /// Sends all pending attachments as separate messages.
-  /// Caption goes on the FIRST attachment so it stays attached visually;
-  /// the rest are bare media markers. Matches Discord / iMessage.
-  Future<void> _sendAttachments(String caption) async {
-    final attachments = List<PendingAttachment>.from(_pendingAttachments);
-    _clearAllPendingAttachments();
-    _messageController.clear();
-    _saveDraftImmediate(widget.conversation.id, '');
-    for (var i = 0; i < attachments.length; i++) {
-      final att = attachments[i];
-      final marker = buildMediaMarker(
-        extension: att.ext,
-        url: att.uploadedUrl!,
-      );
-      await _doSend(marker);
-      if (i == 0 && caption.isNotEmpty) {
-        await _doSend(caption);
-      }
-    }
-    widget.onMessageSent();
-  }
-
-  /// Tries to dispatch [text] as a slash command.
-  /// Returns true if the command was handled (caller should return early).
-  Future<bool> _tryHandleSlashCommand(String text) async {
-    final slashCmd = parseSlashCommand(text);
-    if (slashCmd == null) return false;
-
-    String? rewrittenText;
-    final handled = await dispatchSlashCommand(
-      slashCmd,
-      widget.conversation,
-      ref,
-      context,
-      onRewrite: (newText) {
-        rewrittenText = newText;
-      },
-    );
-    if (!handled) return false;
-
-    _messageController.clear();
-    _saveDraftImmediate(widget.conversation.id, '');
-    // If a fun command rewrote the text, send the rewritten version.
-    if (rewrittenText != null) {
-      await _doSend(rewrittenText!);
-      _dismissPickers();
-      widget.onMessageSent();
-    }
-    return true;
-  }
-
-  /// Clears the text field, saves an empty draft, dismisses pickers,
-  /// and fires the onMessageSent callback.
-  void _clearInputAndNotify() {
-    _messageController.clear();
-    _saveDraftImmediate(widget.conversation.id, '');
-    _dismissPickers();
-    widget.onMessageSent();
-  }
-
-  /// Hides the media/inline pickers if they are currently open.
-  void _dismissPickers() {
-    if (_showMediaPicker) setState(() => _showMediaPicker = false);
-    if (_showInlinePicker) setState(() => _showInlinePicker = false);
-  }
-
-  Future<void> _doSend(String text) async {
-    if (text.isEmpty) return;
-
-    // Expand emoji shortcodes (e.g., :smile: → 😄) at send-time
-    final expandedText = expandEmojiShortcodes(text);
-
-    final conv = widget.conversation;
-    final myUserId = ref.read(authProvider).userId ?? '';
-    final chatState = ref.read(chatProvider);
-    final replyTo = chatState.replyToMessage;
-
-    String peerUserId = '';
-    String? channelId;
-    if (!conv.isGroup) {
-      final peer = conv.members.where((m) => m.userId != myUserId).firstOrNull;
-      peerUserId = peer?.userId ?? '';
-    } else {
-      final channels = ref.read(channelsProvider).channelsFor(conv.id);
-      channelId =
-          widget.selectedTextChannelId ??
-          channels.where((c) => c.isText).firstOrNull?.id;
-    }
-
-    ref
-        .read(chatProvider.notifier)
-        .addOptimistic(
-          peerUserId,
-          expandedText,
-          myUserId,
-          conversationId: conv.id,
-          channelId: channelId,
-          replyToId: replyTo?.id,
-          replyToContent: replyTo?.content,
-          replyToUsername: replyTo?.fromUsername,
-        );
-
-    // Clear reply state after capturing the reply info.
-    if (replyTo != null) {
-      ref.read(chatProvider.notifier).clearReplyTo();
-    }
-
-    try {
-      if (conv.isGroup) {
-        await ref
-            .read(websocketProvider.notifier)
-            .sendGroupMessage(
-              conv.id,
-              expandedText,
-              channelId: channelId,
-              replyToId: replyTo?.id,
-            );
-      } else {
-        await ref
-            .read(websocketProvider.notifier)
-            .sendMessage(
-              peerUserId,
-              expandedText,
-              conversationId: conv.id,
-              replyToId: replyTo?.id,
-            );
-      }
-    } catch (e) {
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to send message',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Input changed / typing indicator / mention detection
-  // ---------------------------------------------------------------------------
-
-  void _onInputChanged(String text) {
-    // Don't send typing indicator while editing an existing message —
-    // recipients would see "X is typing..." when X is only editing.
-    if (!_isEditing) {
-      final conv = widget.conversation;
-      if (text.isNotEmpty) {
-        ref
-            .read(websocketProvider.notifier)
-            .sendTyping(
-              conv.id,
-              channelId: conv.isGroup ? widget.selectedTextChannelId : null,
-            );
-      }
-    }
-    _detectMention(text);
-  }
-
-  void _detectMention(String text) {
-    _mentionController.detect(
-      text: text,
-      cursorPosition: _messageController.selection.baseOffset,
-      isGroup: widget.conversation.isGroup,
-    );
-  }
-
-  void _handleMentionSelected(String username) {
-    final text = _messageController.text;
-    final cursorPos = _messageController.selection.baseOffset;
-
-    // No selection (-1) means the field has never been focused; bail out
-    // before insertMention's no-op path so we don't dismiss the picker
-    // without inserting anything (would be a silent UX failure).
-    if (cursorPos < 0) return;
-
-    _messageController.value = insertMention(
-      text: text,
-      cursorPosition: cursorPos,
-      username: username,
-    );
-
-    _mentionController.dismiss();
-    _inputFocusNode.requestFocus();
-  }
-
-  List<ConversationMember> get _filteredMentionMembers {
-    final myUserId = ref.read(authProvider).userId ?? '';
-    return MentionComposerController.filterMembers(
-      widget.conversation.members,
-      myUserId,
-    );
-  }
-
-  /// Whether the signed-in user has admin or owner role in the current
-  /// conversation.  Always `false` for DMs (non-group conversations).
-  bool get _currentUserIsGroupAdmin {
-    if (!widget.conversation.isGroup) return false;
-    final myUserId = ref.read(authProvider).userId ?? '';
-    final me = widget.conversation.members
-        .where((m) => m.userId == myUserId)
-        .firstOrNull;
-    final role = me?.role?.toLowerCase();
-    return role == 'admin' || role == 'owner';
-  }
-
-  // ---------------------------------------------------------------------------
-  // Pending attachment helpers (Discord-style preview before send)
-  // ---------------------------------------------------------------------------
-
-  /// Stage [bytes] as a new pending attachment and kick off its upload.
-  /// Multiple calls accumulate — the strip shows N chips for N picks until
-  /// the user sends or cancels each.
-  void _setPendingAttachment({
-    required List<int> bytes,
-    required String fileName,
-    required String mimeType,
-    required String ext,
-  }) {
-    final attachment = PendingAttachment(
-      bytes: Uint8List.fromList(bytes),
-      fileName: fileName,
-      mimeType: mimeType,
-      ext: ext,
-      sizeBytes: bytes.length,
-    );
-    setState(() => _pendingAttachments.add(attachment));
-    _startAttachmentUploadFor(attachment);
-  }
-
-  /// Cancel and remove a single staged attachment. Safe to call before the
-  /// upload completes — the [cancelled] flag short-circuits the
-  /// success-path setState in [_startAttachmentUploadFor].
-  void _removePendingAttachment(PendingAttachment attachment) {
-    if (!_pendingAttachments.contains(attachment)) return;
-    attachment.cancelled = true;
-    setState(() => _pendingAttachments.remove(attachment));
-    attachment.dispose();
-  }
-
-  /// Open the [ImageAnnotationEditor] on top of the composer. When the user
-  /// confirms, the original attachment is cancelled and replaced with a new
-  /// pending attachment whose bytes are the annotated PNG (#908). The new
-  /// attachment goes through the regular upload + send path — the
-  /// encrypt/wire flow is untouched.
-  Future<void> _annotatePendingAttachment(PendingAttachment attachment) async {
-    final bytes = attachment.bytes;
-    if (bytes == null) return; // External-URL attachments can't be annotated.
-
-    final navigator = Navigator.of(context);
-    Uint8List? annotated;
-    await navigator.push<void>(
-      MaterialPageRoute<void>(
-        fullscreenDialog: true,
-        builder: (_) => ImageAnnotationEditor(
-          imageBytes: bytes,
-          onConfirm: (Uint8List pngBytes) {
-            annotated = pngBytes;
-            Navigator.of(navigator.context).maybePop();
-          },
-        ),
-      ),
-    );
-    if (!mounted || annotated == null) return;
-
-    // Replace the original chip. Strip the source extension and force PNG
-    // since the rasterised output is always PNG-encoded.
-    final originalName = attachment.fileName;
-    final dot = originalName.lastIndexOf('.');
-    final stem = dot > 0 ? originalName.substring(0, dot) : originalName;
-    final newName = '$stem-annotated.png';
-
-    _removePendingAttachment(attachment);
-    _setPendingAttachment(
-      bytes: annotated!,
-      fileName: newName,
-      mimeType: 'image/png',
-      ext: 'png',
-    );
-  }
-
-  /// Stage an external-URL attachment (e.g. picked from the GIF browser).
-  /// No upload is performed — the URL is used as-is on send.
-  void _setPendingExternalAttachment({
-    required String url,
-    required String fileName,
-    required String mimeType,
-    required String ext,
-  }) {
-    final attachment = PendingAttachment(
-      bytes: null,
-      fileName: fileName,
-      mimeType: mimeType,
-      ext: ext,
-      sizeBytes: 0,
-      uploadedUrl: url,
-    );
-    setState(() => _pendingAttachments.add(attachment));
-  }
-
-  /// Drop all staged attachments. Called after a successful send.
-  void _clearAllPendingAttachments() {
-    if (_pendingAttachments.isEmpty) return;
-    final toDispose = List<PendingAttachment>.from(_pendingAttachments);
-    setState(() => _pendingAttachments.clear());
-    for (final att in toDispose) {
-      att.dispose();
-    }
-  }
-
-  Future<void> _startAttachmentUploadFor(PendingAttachment att) async {
-    if (att.isExternalUrl) return; // already has a URL (e.g. GIF picker).
-    final bytes = att.bytes!;
-    final serverUrl = ref.read(serverUrlProvider);
-    try {
-      final result = await _uploadWithAuthRetry(
-        serverUrl: serverUrl,
-        bytes: bytes,
-        fileName: att.fileName,
-        mimeType: att.mimeType,
-        onProgress: (sent, total) {
-          if (att.cancelled || total <= 0) return;
-          att.progress.value = (sent / total).clamp(0.0, 1.0);
-        },
-      );
-      if (!mounted || att.cancelled) return;
-
-      if (result != null) {
-        setState(() {
-          att.uploadedUrl = result;
-          att.isUploading = false;
-          att.progress.value = 1.0;
-        });
-        _inputFocusNode.requestFocus();
-      } else {
-        if (mounted) {
-          ToastService.show(
-            context,
-            'Upload failed: ${att.fileName}',
-            type: ToastType.error,
-          );
-        }
-        _removePendingAttachment(att);
-      }
-    } catch (e) {
-      if (!mounted || att.cancelled) return;
-      ToastService.show(
-        context,
-        'Upload failed: ${att.fileName}',
-        type: ToastType.error,
-      );
-      _removePendingAttachment(att);
-    }
-  }
-
-  /// Upload media with automatic 401→token-refresh→retry via [UploadClient].
-  ///
-  /// Returns the server-assigned URL on success, or null on failure.
-  Future<String?> _uploadWithAuthRetry({
-    required String serverUrl,
-    required List<int> bytes,
-    required String fileName,
-    required String mimeType,
-    void Function(int sent, int total)? onProgress,
-  }) async {
-    // If the user has the "preserve original filenames" privacy toggle off,
-    // upload the file under a generic name keyed off the extension. The file
-    // contents are unchanged.
-    final preserve = await readPreserveOriginalFilenames();
-    final uploadFileName = preserve ? fileName : genericFilename(fileName);
-
-    final uploader = UploadClient(ref.read(authProvider.notifier));
-    final result = await uploader.uploadFile(
-      serverUrl: serverUrl,
-      path: '/api/media/upload',
-      bytes: bytes,
-      fileName: uploadFileName,
-      mimeType: mimeType,
-      extraFields: {'conversation_id': widget.conversation.id},
-      onProgress: onProgress,
-    );
-
-    if (result.ok) {
-      // Pre-populate the dimension cache so ImageAttachment can reserve the
-      // correct aspect-ratio placeholder on first render, before bytes arrive.
-      final url = result.url;
-      final w = result.width;
-      final h = result.height;
-      if (url != null && w != null && h != null) {
-        cacheImageDimensions(url, w, h);
-      }
-      return url;
-    }
-
-    if (mounted && !result.ok) {
-      ToastService.show(
-        context,
-        'Upload failed (${result.statusCode})',
-        type: ToastType.error,
-      );
-    }
-    return null;
-  }
-
-  Future<void> _pasteImageFromClipboard() async {
-    final image = await readImageFromClipboard();
-    if (image == null) return;
-    if (!mounted) return;
-
-    _setPendingAttachment(
-      bytes: image.bytes,
-      fileName: image.fileName,
-      mimeType: image.mimeType,
-      ext: extensionFromMime(image.mimeType),
-    );
-  }
-
-  /// Handle Ctrl+V: read clipboard text and insert at cursor, bypassing the
-  /// browser's native paste context menu that CanvasKit shows by default.
-  /// Also tries image paste in parallel for clipboard images.
-  Future<void> _handlePaste() async {
-    // Try image paste (non-blocking)
-    if (!_isEditing) {
-      _pasteImageFromClipboard();
-    }
-
-    // Read text from clipboard and insert at cursor
-    final data = await Clipboard.getData('text/plain');
-    if (data?.text == null || data!.text!.isEmpty) return;
-    if (!mounted) return;
-
-    final text = data.text!;
-    final selection = _messageController.selection;
-    final currentText = _messageController.text;
-
-    if (selection.isValid) {
-      final newText = currentText.replaceRange(
-        selection.start,
-        selection.end,
-        text,
-      );
-      _messageController.value = TextEditingValue(
-        text: newText,
-        selection: TextSelection.collapsed(
-          offset: selection.start + text.length,
-        ),
-      );
-    } else {
-      _messageController.value = TextEditingValue(
-        text: currentText + text,
-        selection: TextSelection.collapsed(
-          offset: currentText.length + text.length,
-        ),
-      );
-    }
-
-    _onInputChanged(_messageController.text);
-  }
-
-  /// Upload [bytes] and immediately send the result as a message.
-  /// Used for the 2nd..Nth files when multiple are selected at once.
-  Future<void> _sendFileImmediately({
-    required Uint8List bytes,
-    required String fileName,
-    required String mimeType,
-    required String ext,
-  }) async {
-    final serverUrl = ref.read(serverUrlProvider);
-    final url = await _uploadWithAuthRetry(
-      serverUrl: serverUrl,
-      bytes: bytes,
-      fileName: fileName,
-      mimeType: mimeType,
-    );
-    if (!mounted || url == null) return;
-    final marker = buildMediaMarker(extension: ext, url: url);
-    await _doSend(marker);
-  }
-
-  Future<void> _pickFile() => pickers.pickFile(
-    context: context,
-    mounted: () => mounted,
-    isPicking: () => _isPickingFile,
-    setIsPicking: (v) => _isPickingFile = v,
-    stage:
-        ({
-          required List<int> bytes,
-          required String fileName,
-          required String mimeType,
-          required String ext,
-        }) => _setPendingAttachment(
-          bytes: bytes,
-          fileName: fileName,
-          mimeType: mimeType,
-          ext: ext,
-        ),
-    sendImmediately: _sendFileImmediately,
-  );
-
-  // ---------------------------------------------------------------------------
-  // Edit mode
-  // ---------------------------------------------------------------------------
-
-  void _cancelEditMode() {
-    _draftSaveTimer?.cancel();
-    _suppressDraftSave = true;
-    setState(() {
-      _controller.exitEditMode();
-      _messageController.clear();
-      _isTextEmpty = true;
-    });
-    // Restore the saved draft (if any) after leaving edit mode.
-    _loadDraft(widget.conversation.id).then((_) {
-      _suppressDraftSave = false;
-    });
-  }
-
-  Future<void> _submitEdit() async {
-    final text = _messageController.text.trim();
-    if (text.isEmpty || _editingMessage == null) return;
-
-    final conv = widget.conversation;
-    // #582: belt-and-suspenders for the parent UI gate. Editing on an
-    // encrypted conversation would broadcast plaintext to every member, so
-    // we never submit it. The server also returns 409.
-    if (conv.isEncrypted) {
-      _cancelEditMode();
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Edit unsupported for encrypted messages.',
-          type: ToastType.info,
-        );
-      }
-      return;
-    }
-    final messageId = _editingMessage!.id;
-    final serverUrl = ref.read(serverUrlProvider);
-
-    // Optimistically update local state
-    ref.read(chatProvider.notifier).editMessage(conv.id, messageId, text);
-    _cancelEditMode();
-
-    try {
-      final response = await ref
-          .read(authProvider.notifier)
-          .authenticatedRequest(
-            (token) => http.put(
-              Uri.parse('$serverUrl/api/messages/$messageId'),
-              headers: {
-                'Authorization': 'Bearer $token',
-                'Content-Type': 'application/json',
-              },
-              body: jsonEncode({'content': text}),
-            ),
-          );
-      // The server returns 409 when an encrypted conversation rejects an
-      // edit (#582). Surface a non-fatal toast so the user understands
-      // why the change rolled back.
-      if (response.statusCode == 409 && mounted) {
-        ToastService.show(
-          context,
-          'Edit unsupported for encrypted messages.',
-          type: ToastType.info,
-        );
-      } else if (response.statusCode >= 400 && mounted) {
-        ToastService.show(
-          context,
-          'Failed to edit message',
-          type: ToastType.error,
-        );
-      }
-    } catch (e) {
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to edit message',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Voice state sync (for push-to-talk key handling)
-  // ---------------------------------------------------------------------------
-
-  Future<void> _syncVoiceState() async {
-    final conv = widget.conversation;
-    final channelId = widget.effectiveActiveVoiceChannelId;
-    if (channelId == null) return;
-    final voiceSettings = ref.read(voiceSettingsProvider);
-    await ref
-        .read(channelsProvider.notifier)
-        .updateVoiceState(
-          conversationId: conv.id,
-          channelId: channelId,
-          isMuted: voiceSettings.selfMuted,
-          isDeafened: voiceSettings.selfDeafened,
-          pushToTalk: voiceSettings.pushToTalkEnabled,
-        );
-  }
-
-  // ---------------------------------------------------------------------------
-  // Voice recording
-  // ---------------------------------------------------------------------------
-
-  Future<void> _startRecording() async {
-    if (kIsWeb) {
-      // Web recording not supported via the record package in this config.
-      ToastService.show(
-        context,
-        'Voice messages are not supported in the browser yet',
-        type: ToastType.info,
-      );
-      return;
-    }
-
-    final hasPermission = await _recorder.hasPermission();
-    if (!hasPermission) {
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Microphone permission is required to send voice messages',
-          type: ToastType.error,
-        );
-      }
-      return;
-    }
-
-    final dir = await getTemporaryDirectory();
-    final path =
-        '${dir.path}/voice_${DateTime.now().millisecondsSinceEpoch}.m4a';
-
-    try {
-      await _recorder.start(
-        const RecordConfig(
-          encoder: AudioEncoder.aacLc,
-          sampleRate: 44100,
-          bitRate: 64000,
-          numChannels: 1,
-        ),
-        path: path,
-      );
-    } catch (e) {
-      // record's iOS / Android backends can throw on session-init failures
-      // (audio session in use by another app, codec unavailable, etc.).
-      // Without surfacing this, the user just saw nothing happen when they
-      // tried to record (#554).
-      debugPrint('[ChatInput] _recorder.start failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Could not start recording: $e',
-          type: ToastType.error,
-        );
-      }
-      return;
-    }
-
-    setState(() {
-      _isRecording = true;
-      _recordingStartTime = DateTime.now();
-      _recordingDuration = Duration.zero;
-      _recordingAmplitudes.clear();
-    });
-
-    // Tick every 100ms to update the duration counter and collect amplitudes.
-    _recordingTimer = Timer.periodic(const Duration(milliseconds: 100), (
-      _,
-    ) async {
-      if (!_isRecording || !mounted) return;
-      final elapsed = DateTime.now().difference(_recordingStartTime!);
-      setState(() => _recordingDuration = elapsed);
-
-      try {
-        final amp = await _recorder.getAmplitude();
-        // amp.current is in dBFS [-160, 0]. Map to [0, 1].
-        final normalised = ((amp.current + 60) / 60).clamp(0.0, 1.0);
-        _recordingAmplitudes.add(normalised);
-      } catch (_) {
-        // Amplitude polling is best-effort.
-      }
-    });
-  }
-
-  Future<void> _stopRecording({bool cancel = false}) async {
-    _recordingTimer?.cancel();
-    _recordingTimer = null;
-
-    if (!_isRecording) return;
-
-    final path = await _recorder.stop();
-
-    setState(() {
-      _isRecording = false;
-      _recordingDuration = Duration.zero;
-    });
-
-    if (cancel || path == null) {
-      _recordingAmplitudes.clear();
-      return;
-    }
-
-    // Read the recorded bytes and attach as a pending voice message.
-    try {
-      final file = File(path);
-      if (!file.existsSync()) {
-        if (mounted) {
-          ToastService.show(
-            context,
-            'Recording was lost — no audio file produced',
-            type: ToastType.error,
-          );
-        }
-        return;
-      }
-      final bytes = await file.readAsBytes();
-      _recordingAmplitudes.clear();
-
-      // Reject empty / near-empty recordings rather than uploading a
-      // 0-byte file that the server will reject anyway. The threshold is
-      // generous — even a 100ms aac frame is hundreds of bytes (#554).
-      if (bytes.length < 256) {
-        if (mounted) {
-          ToastService.show(
-            context,
-            'Recording was too short or silent (${bytes.length}B)',
-            type: ToastType.error,
-          );
-        }
-        return;
-      }
-
-      _setPendingVoiceAttachment(bytes: bytes);
-    } catch (e) {
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Could not read recording: $e',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  void _setPendingVoiceAttachment({required Uint8List bytes}) {
-    final fileName = 'voice_${DateTime.now().millisecondsSinceEpoch}.m4a';
-    _setPendingAttachment(
-      bytes: bytes,
-      fileName: fileName,
-      mimeType: 'audio/mp4',
-      ext: 'm4a',
-    );
-  }
-
-  // ---------------------------------------------------------------------------
-  // Build helpers -- kept in root widget
-  // ---------------------------------------------------------------------------
-
-  void _showMobileAttachMenu() {
-    showEchoBottomSheet<void>(
-      context,
-      builder: (ctx) => Padding(
-        padding: const EdgeInsets.symmetric(vertical: 16),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            AttachOption(
-              icon: Icons.photo_library_outlined,
-              label: 'Photos & Videos',
-              onTap: () {
-                Navigator.pop(ctx);
-                _pickImageFromGallery();
-              },
-            ),
-            AttachOption(
-              icon: Icons.camera_alt_outlined,
-              label: 'Camera',
-              onTap: () {
-                Navigator.pop(ctx);
-                _pickImageFromCamera();
-              },
-            ),
-            AttachOption(
-              icon: Icons.insert_drive_file_outlined,
-              label: 'File',
-              onTap: () {
-                Navigator.pop(ctx);
-                _pickFile();
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Future<void> _pickImageFromGallery() => pickers.pickImageFromGallery(
-    context: context,
-    mounted: () => mounted,
-    isPicking: () => _isPickingFile,
-    setIsPicking: (v) => _isPickingFile = v,
-    stage:
-        ({
-          required List<int> bytes,
-          required String fileName,
-          required String mimeType,
-          required String ext,
-        }) => _setPendingAttachment(
-          bytes: bytes,
-          fileName: fileName,
-          mimeType: mimeType,
-          ext: ext,
-        ),
-    sendImmediately: _sendFileImmediately,
-  );
-
-  Future<void> _pickImageFromCamera() => pickers.pickImageFromCamera(
-    context: context,
-    mounted: () => mounted,
-    isPicking: () => _isPickingFile,
-    setIsPicking: (v) => _isPickingFile = v,
-    stage:
-        ({
-          required List<int> bytes,
-          required String fileName,
-          required String mimeType,
-          required String ext,
-        }) => _setPendingAttachment(
-          bytes: bytes,
-          fileName: fileName,
-          mimeType: mimeType,
-          ext: ext,
-        ),
-  );
-
-  /// Toggles the inline (mobile) or overlay (desktop) media picker. Shared
-  /// callback wired into [MediaPickerToggle] from `_buildInputRow`.
-  void _toggleMediaPicker(bool isMobileLayout) {
-    if (isMobileLayout) {
-      if (_showInlinePicker) {
-        setState(() => _showInlinePicker = false);
-        _inputFocusNode.requestFocus();
-      } else {
-        _inputFocusNode.unfocus();
-        Future<void>.delayed(const Duration(milliseconds: 300), () {
-          if (mounted) setState(() => _showInlinePicker = true);
-        });
-      }
-      widget.onMediaPickerChanged?.call();
-    } else {
-      setState(() => _showMediaPicker = !_showMediaPicker);
-      widget.onMediaPickerChanged?.call();
-      if (!_showMediaPicker) {
-        _inputFocusNode.requestFocus();
-      }
-    }
-  }
-
-  /// Expose inline picker state for ChatPanel layout adjustments.
-  bool get showInlinePicker => _showInlinePicker;
-
-  /// Handles push-to-talk key events. Returns true if the event was consumed.
-  bool _handlePushToTalk(
-    KeyEvent event,
-    VoiceSettingsState voiceSettings,
-    String? effectiveActiveVoiceChannelId,
-  ) {
-    final pttKeyId = voiceSettings.pushToTalkKeyId;
-    final isPttKey = event.logicalKey.keyId.toString() == pttKeyId;
-    final canPushToTalk =
-        voiceSettings.pushToTalkEnabled &&
-        effectiveActiveVoiceChannelId != null;
-
-    if (!canPushToTalk || !isPttKey) return false;
-
-    final allowCapture =
-        !voiceSettings.selfMuted && !voiceSettings.selfDeafened;
-    if (event is KeyDownEvent && allowCapture) {
-      ref.read(voiceRtcProvider.notifier).setCaptureEnabled(true);
-      _syncVoiceState();
-    } else if (event is KeyUpEvent) {
-      ref.read(voiceRtcProvider.notifier).setCaptureEnabled(false);
-      _syncVoiceState();
-    }
-    return false; // don't consume -- let other handlers also run
-  }
-
-  /// Handles the Escape key by dismissing the highest-priority
-  /// composer-modal state (mention picker → inline picker → media
-  /// picker → pending attachments → edit mode → reply state).  Each
-  /// dismissal is a no-op if the corresponding state isn't active.
-  void _handleEscapeKey() {
-    if (_mentionController.showPicker) {
-      _mentionController.dismiss();
-    } else if (_showInlinePicker) {
-      setState(() => _showInlinePicker = false);
-      _inputFocusNode.requestFocus();
-    } else if (_showMediaPicker) {
-      setState(() => _showMediaPicker = false);
-      _inputFocusNode.requestFocus();
-    } else if (_hasPendingAttachment) {
-      _clearAllPendingAttachments();
-    } else if (_isEditing) {
-      _cancelEditMode();
-    } else if (ref.read(chatProvider).replyToMessage != null) {
-      ref.read(chatProvider.notifier).clearReplyTo();
-    }
-  }
-
-  /// Handles Ctrl+V paste shortcut. Returns a [KeyEventResult] indicating
-  /// whether the event was consumed.
-  ///
-  /// Previously returned `ignored` on web and Linux, which meant
-  /// `_handlePaste()` (including image paste) was never called on those
-  /// platforms. Now always invokes `_handlePaste()` and marks handled.
-  KeyEventResult _handlePasteShortcut() {
-    _handlePaste();
-    return KeyEventResult.handled;
-  }
-
-  /// Handles Ctrl+C / Ctrl+X copy/cut shortcuts. Returns a [KeyEventResult]
-  /// indicating whether the event was consumed.
-  KeyEventResult _handleCopyCutShortcut(bool isCut) {
-    if (kIsWeb || defaultTargetPlatform == TargetPlatform.linux) {
-      return KeyEventResult.ignored;
-    }
-    final sel = _messageController.selection;
-    if (sel.isValid && !sel.isCollapsed) {
-      final selected = _messageController.text.substring(sel.start, sel.end);
-      Clipboard.setData(ClipboardData(text: selected));
-      if (isCut) {
-        _messageController.text = _messageController.text.replaceRange(
-          sel.start,
-          sel.end,
-          '',
-        );
-        _messageController.selection = TextSelection.collapsed(
-          offset: sel.start,
-        );
-      }
-      return KeyEventResult.handled;
-    }
-    return KeyEventResult.ignored;
-  }
-
-  /// Handle Ctrl/Cmd key shortcuts (paste, copy, cut).
-  KeyEventResult _handleModifierShortcut(KeyEvent event) {
-    final isCtrlOrMeta =
-        HardwareKeyboard.instance.isControlPressed ||
-        HardwareKeyboard.instance.isMetaPressed;
-    if (!isCtrlOrMeta) return KeyEventResult.ignored;
-
-    if (event.logicalKey == LogicalKeyboardKey.keyV) {
-      return _handlePasteShortcut();
-    }
-    if (event.logicalKey == LogicalKeyboardKey.keyC) {
-      return _handleCopyCutShortcut(false);
-    }
-    if (event.logicalKey == LogicalKeyboardKey.keyX) {
-      // Ctrl+Shift+X = strikethrough; plain Ctrl+X = cut.
-      if (HardwareKeyboard.instance.isShiftPressed) {
-        return _applyMarkdownWrap(open: '~~', close: '~~');
-      }
-      return _handleCopyCutShortcut(true);
-    }
-    if (event.logicalKey == LogicalKeyboardKey.keyB &&
-        !HardwareKeyboard.instance.isShiftPressed) {
-      return _applyMarkdownWrap(open: '**', close: '**');
-    }
-    if (event.logicalKey == LogicalKeyboardKey.keyI &&
-        !HardwareKeyboard.instance.isShiftPressed) {
-      return _applyMarkdownWrap(open: '*', close: '*');
-    }
-    return KeyEventResult.ignored;
-  }
-
-  /// Wrap the current selection with [open]/[close] markers. When nothing is
-  /// selected, insert both markers at the cursor and leave the caret between
-  /// them so the user can immediately type. Returns [KeyEventResult.handled]
-  /// so the key does not also reach the underlying TextField.
-  KeyEventResult _applyMarkdownWrap({
-    required String open,
-    required String close,
-  }) {
-    final text = _messageController.text;
-    final sel = _messageController.selection;
-    if (!sel.isValid) return KeyEventResult.ignored;
-
-    if (sel.isCollapsed) {
-      final cursor = sel.start;
-      final newText =
-          text.substring(0, cursor) + open + close + text.substring(cursor);
-      _messageController.value = TextEditingValue(
-        text: newText,
-        selection: TextSelection.collapsed(offset: cursor + open.length),
-      );
-      return KeyEventResult.handled;
-    }
-
-    final before = text.substring(0, sel.start);
-    final selected = text.substring(sel.start, sel.end);
-    final after = text.substring(sel.end);
-    final newText = '$before$open$selected$close$after';
-    _messageController.value = TextEditingValue(
-      text: newText,
-      selection: TextSelection(
-        baseOffset: sel.start + open.length,
-        extentOffset: sel.end + open.length,
-      ),
-    );
-    return KeyEventResult.handled;
-  }
-
-  /// Up arrow with empty input: edit last own message (Discord behavior).
-  /// #582: skip on encrypted conversations to avoid surfacing an edit flow
-  /// the server will reject.
-  KeyEventResult _handleArrowUpEditLast() {
-    if (!_isTextEmpty || _isEditing) return KeyEventResult.ignored;
-    if (widget.conversation.isEncrypted) return KeyEventResult.ignored;
-    final messages = ref
-        .read(chatProvider)
-        .messagesForConversation(widget.conversation.id);
-    final myUserId = ref.read(authProvider).userId ?? '';
-    final lastOwn = messages
-        .where((m) => m.isMine && m.fromUserId == myUserId)
-        .lastOrNull;
-    if (lastOwn != null) {
-      enterEditMode(lastOwn);
-      return KeyEventResult.handled;
-    }
-    return KeyEventResult.ignored;
-  }
-
-  KeyEventResult _onKeyEvent(
-    FocusNode _,
-    KeyEvent event,
-    VoiceSettingsState voiceSettings,
-    String? effectiveActiveVoiceChannelId,
-  ) {
-    _handlePushToTalk(event, voiceSettings, effectiveActiveVoiceChannelId);
-
-    if (event is! KeyDownEvent) return KeyEventResult.ignored;
-
-    final mentionResult = _handleMentionPickerIfActive(event);
-    if (mentionResult != KeyEventResult.ignored) return mentionResult;
-
-    if (event.logicalKey == LogicalKeyboardKey.escape) {
-      _handleEscapeKey();
-    }
-
-    final submitResult = _handleEnterSubmit(event);
-    if (submitResult != KeyEventResult.ignored) return submitResult;
-
-    final modResult = _handleModifierShortcut(event);
-    if (modResult != KeyEventResult.ignored) return modResult;
-
-    if (event.logicalKey == LogicalKeyboardKey.arrowUp) {
-      return _handleArrowUpEditLast();
-    }
-
-    return KeyEventResult.ignored;
-  }
-
-  // Mention picker keyboard nav takes precedence — Tab/Enter accept the
-  // selected row, arrows move it, Escape closes the picker (without
-  // also bubbling Escape through to message edit-cancel). Space falls
-  // through — extractMentionQuery sees the space and closes the picker,
-  // leaving the literal "@" in the text.
-  KeyEventResult _handleMentionPickerIfActive(KeyDownEvent event) {
-    if (!_mentionController.showPicker) return KeyEventResult.ignored;
-    return _handleMentionPickerKey(event);
-  }
-
-  KeyEventResult _handleEnterSubmit(KeyDownEvent event) {
-    if (event.logicalKey == LogicalKeyboardKey.enter &&
-        !HardwareKeyboard.instance.isShiftPressed) {
-      _isEditing ? _submitEdit() : _sendMessage();
-      return KeyEventResult.handled;
-    }
-    return KeyEventResult.ignored;
-  }
-
-  /// Keyboard handler for the mention picker overlay. Returns
-  /// [KeyEventResult.ignored] when the picker should *not* consume the
-  /// event (e.g. Space, so the picker auto-dismisses via the query
-  /// extractor).
-  KeyEventResult _handleMentionPickerKey(KeyDownEvent event) {
-    if (event.logicalKey == LogicalKeyboardKey.tab ||
-        (event.logicalKey == LogicalKeyboardKey.enter &&
-            !HardwareKeyboard.instance.isShiftPressed)) {
-      _acceptMentionSelection();
-      return KeyEventResult.handled;
-    }
-    if (event.logicalKey == LogicalKeyboardKey.arrowDown) {
-      _moveMentionSelection(_MentionMove.down);
-      return KeyEventResult.handled;
-    }
-    if (event.logicalKey == LogicalKeyboardKey.arrowUp) {
-      _moveMentionSelection(_MentionMove.up);
-      return KeyEventResult.handled;
-    }
-    if (event.logicalKey == LogicalKeyboardKey.escape) {
-      _mentionController.dismiss();
-      return KeyEventResult.handled;
-    }
-    return KeyEventResult.ignored;
-  }
-
-  Widget _buildTextField({
-    required bool showMediaPicker,
-    required VoiceSettingsState voiceSettings,
-    required String? effectiveActiveVoiceChannelId,
-  }) {
-    return Expanded(
-      child: Focus(
-        onKeyEvent: (node, event) => _onKeyEvent(
-          node,
-          event,
-          voiceSettings,
-          effectiveActiveVoiceChannelId,
-        ),
-        child: TextField(
-          controller: _messageController,
-          focusNode: _inputFocusNode,
-          maxLines: 10,
-          minLines: 1,
-          maxLength: 4000,
-          buildCounter:
-              (
-                context, {
-                required currentLength,
-                required isFocused,
-                required maxLength,
-              }) {
-                if (currentLength <= 3200) return null;
-                final counterColor = currentLength > 3900
-                    ? EchoTheme.danger
-                    : context.textMuted;
-                return Text(
-                  '$currentLength/$maxLength',
-                  style: TextStyle(color: counterColor, fontSize: 11),
-                );
-              },
-          autofillHints: const [],
-          style: TextStyle(fontSize: 14, color: context.textPrimary),
-          decoration: InputDecoration(
-            hintText: _isEditing ? 'Edit your message…' : 'Message — encrypted',
-            border: InputBorder.none,
-            enabledBorder: InputBorder.none,
-            focusedBorder: InputBorder.none,
-            filled: false,
-            contentPadding: const EdgeInsets.symmetric(vertical: 10),
-          ),
-          onChanged: _onInputChanged,
-          onTap: () {
-            if (_showInlinePicker) {
-              setState(() => _showInlinePicker = false);
-            }
-            if (showMediaPicker) {
-              setState(() => _showMediaPicker = false);
-            }
-          },
-          onSubmitted: (_) => _isEditing ? _submitEdit() : _sendMessage(),
-        ),
-      ),
-    );
-  }
-
-  VoidCallback _resolvedSendAction() {
-    if (_isEditing) return _submitEdit;
-    return _sendMessage;
-  }
-
-  // ---------------------------------------------------------------------------
-  // Markdown toolbar
-  // ---------------------------------------------------------------------------
-
-  Future<void> _showLinkDialog() async {
-    final urlController = TextEditingController();
-    final url = await showDialog<String>(
-      context: context,
-      builder: (ctx) => AlertDialog(
-        title: const Text('Insert link'),
-        content: TextField(
-          controller: urlController,
-          autofocus: true,
-          keyboardType: TextInputType.url,
-          decoration: const InputDecoration(hintText: 'https://…'),
-          onSubmitted: (v) => Navigator.pop(ctx, v.trim()),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(ctx),
-            child: const Text('Cancel'),
-          ),
-          TextButton(
-            onPressed: () => Navigator.pop(ctx, urlController.text.trim()),
-            child: const Text('Insert'),
-          ),
-        ],
-      ),
-    );
-    urlController.dispose();
-    if (url == null || url.isEmpty) return;
-    if (!mounted) return;
-
-    final sel = _messageController.selection;
-    final text = _messageController.text;
-    final hasSelection = sel.isValid && sel.start != sel.end;
-    final label = hasSelection ? text.substring(sel.start, sel.end) : 'link';
-
-    if (hasSelection) {
-      // Replace selection with [label](url)
-      final newText =
-          '${text.substring(0, sel.start)}[$label]($url)${text.substring(sel.end)}';
-      _messageController.value = TextEditingValue(
-        text: newText,
-        selection: TextSelection.collapsed(
-          offset: sel.start + newText.length - text.substring(sel.end).length,
-        ),
-      );
-    } else {
-      // Insert at cursor
-      final pos = sel.isValid ? sel.start : text.length;
-      final inserted = '[$label]($url)';
-      final newText = text.substring(0, pos) + inserted + text.substring(pos);
-      _messageController.value = TextEditingValue(
-        text: newText,
-        selection: TextSelection.collapsed(offset: pos + inserted.length),
-      );
-    }
-    _inputFocusNode.requestFocus();
-  }
-
-  Widget _buildInputRow({
-    required bool showMediaPicker,
-    required bool isMobileLayout,
-    required VoiceSettingsState voiceSettings,
-    required String? effectiveActiveVoiceChannelId,
-  }) {
-    // While recording, replace the entire input row with the recording UI.
-    if (_isRecording) {
-      return RecordingRow(
-        recordingDuration: _recordingDuration,
-        recordingAmplitudes: _recordingAmplitudes,
-        onCancel: () => _stopRecording(cancel: true),
-        onStop: () => _stopRecording(),
-      );
-    }
-
-    // Three-element design: bordered round + button on the left, pill input
-    // (with emoji glyph inside on the right), bordered round mic/send on
-    // the right. Edit mode tints the pill border accent.
-    final pillBorderColor = _isEditing ? context.accent : context.border;
-
-    return Row(
-      crossAxisAlignment: _isMultiline
-          ? CrossAxisAlignment.end
-          : CrossAxisAlignment.center,
-      children: [
-        if (!_isEditing)
-          AttachFileButton(
-            onPickFile: _pickFile,
-            onShowMobileMenu: _showMobileAttachMenu,
-          ),
-        if (!_isEditing) const SizedBox(width: 8),
-        Expanded(
-          child: Container(
-            constraints: const BoxConstraints(minHeight: 40),
-            padding: const EdgeInsets.only(left: 12, right: 4),
-            decoration: BoxDecoration(
-              color: context.surface,
-              borderRadius: BorderRadius.circular(22),
-              border: Border.all(color: pillBorderColor, width: 1),
-            ),
-            child: Row(
-              crossAxisAlignment: _isMultiline
-                  ? CrossAxisAlignment.end
-                  : CrossAxisAlignment.center,
-              children: [
-                _buildTextField(
-                  showMediaPicker: showMediaPicker,
-                  voiceSettings: voiceSettings,
-                  effectiveActiveVoiceChannelId: effectiveActiveVoiceChannelId,
-                ),
-                MediaPickerToggle(
-                  showMediaPicker: showMediaPicker,
-                  onToggle: () => _toggleMediaPicker(isMobileLayout),
-                ),
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(width: 8),
-        SendButton(
-          isTextEmpty: _isTextEmpty,
-          allPendingAttachmentsReady: _allPendingAttachmentsReady,
-          isEditing: _isEditing,
-          isDm: !widget.conversation.isGroup,
-          onStartRecording: _startRecording,
-          resolveSendAction: _resolvedSendAction,
-        ),
-      ],
-    );
-  }
-
   /// Whether the media picker is currently shown. Exposed for ChatPanel
   /// to render the picker in its own overlay Stack.
   bool get showMediaPicker => _showMediaPicker;
+
+  /// Expose inline picker state for ChatPanel layout adjustments.
+  bool get showInlinePicker => _showInlinePicker;
 
   /// Build the media picker panel. Called by ChatPanel to render it
   /// above the message list (not inside the input bar's Stack).
@@ -1784,21 +343,6 @@ class ChatInputBarState extends ConsumerState<ChatInputBar> {
         setState(() => _showMediaPicker = false);
         _inputFocusNode.requestFocus();
       },
-    );
-  }
-
-  /// Inserts [emojiStr] at the current cursor position in the text field.
-  void _insertEmoji(String emojiStr) {
-    final text = _messageController.text;
-    final selection = _messageController.selection;
-    final cursorPos = selection.baseOffset >= 0
-        ? selection.baseOffset
-        : text.length;
-    final newText =
-        text.substring(0, cursorPos) + emojiStr + text.substring(cursorPos);
-    _messageController.text = newText;
-    _messageController.selection = TextSelection.collapsed(
-      offset: cursorPos + emojiStr.length,
     );
   }
 
@@ -1916,83 +460,6 @@ class ChatInputBarState extends ConsumerState<ChatInputBar> {
         ),
         // Media picker is rendered by ChatPanel in its own Stack (above message list).
       ],
-    );
-  }
-
-  /// Builds the mention and slash-command autocomplete pickers.
-  /// Exactly one is visible at a time — mention picker takes priority.
-  Widget _buildAutocompletePickers() {
-    if (_mentionController.showPicker) {
-      return MentionAutocomplete(
-        members: _filteredMentionMembers,
-        mentionQuery: _mentionController.query,
-        selectedIndex: _mentionPickerIndex,
-        onMentionSelected: _handleMentionSelected,
-      );
-    }
-    return SlashCommandAutocomplete(
-      inputText: _messageController.text,
-      userIsGroupAdmin: _currentUserIsGroupAdmin,
-      onSelect: (template) {
-        _messageController.text = template;
-        _messageController.selection = TextSelection.collapsed(
-          offset: template.length,
-        );
-      },
-    );
-  }
-
-  /// Builds the animated inline media/emoji picker that replaces the keyboard
-  /// on mobile. Collapses to [SizedBox.shrink] when not shown.
-  Widget _buildInlinePicker() {
-    final pickerHeight = (_lastKeyboardHeight > 0 ? _lastKeyboardHeight : 280)
-        .toDouble();
-    return AnimatedSize(
-      duration: const Duration(milliseconds: 220),
-      curve: Curves.easeInOut,
-      child: _showInlinePicker
-          ? SizedBox(
-              height: pickerHeight,
-              child: MobileMediaPickerPanel(
-                onEmojiSelected: (category, emoji) {
-                  _insertEmoji(emoji.emoji);
-                  _onInputChanged(_messageController.text);
-                },
-                onGifSelected: (gifUrl, slug) {
-                  setState(() => _showInlinePicker = false);
-                  _setPendingExternalAttachment(
-                    url: gifUrl,
-                    fileName: 'gif',
-                    mimeType: kImageGifMimeType,
-                    ext: 'gif',
-                  );
-                },
-                onPhotoSelected: _handlePhotoSelected,
-                onClose: () {
-                  setState(() => _showInlinePicker = false);
-                  _inputFocusNode.requestFocus();
-                },
-              ),
-            )
-          : const SizedBox.shrink(),
-    );
-  }
-
-  /// Handle a photo selected from the camera roll gallery.
-  Future<void> _handlePhotoSelected(
-    File file,
-    String fileName,
-    String mimeType,
-  ) async {
-    final bytes = await file.readAsBytes();
-    if (!mounted) return;
-    final ext = fileName.split('.').last.toLowerCase();
-    setState(() => _showInlinePicker = false);
-    _setPendingAttachment(
-      bytes: bytes,
-      fileName: fileName,
-      mimeType: mimeType,
-      ext: ext,
     );
   }
 }
